add 19

+++ auto_load_data_rnn.py
@@ -94,8 +94,8 @@ test_data = keras.preprocessing.sequence.pad_sequences(test_data,
                                                        maxlen=256)
 
 def pre_build():
-    input = tf.placeholder(tf.int32, [None, None], "input")
-    targets = tf.placeholder(tf.int32, [None, None], "target")
+    input = tf.placeholder(tf.float32, [None, None], "input")
+    targets = tf.placeholder(tf.float32, [None, None], "target")
     keep_prob = tf.placeholder(tf.float32, name='keep_prob')
 
     return input, targets, keep_prob
